requirements.txt
boto3
openai>=1.0.0
python-dotenv
jinja2
